requirements.txt
python-dotenv==1.0.0
python-telegram-bot==20.1
requests==2.28.2
